Move HierarchicalDataCommunicatorConstants to the correct package
