+++ shared/src/main/java/com/vaadin/shared/data/HierarchicalDataCommunicatorConstants.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.shared.ui.treegrid;
+package com.vaadin.shared.data;
 
 import java.io.Serializable;
